fix: parameterize SQL in sequence queries, batch operations, and CLI
Cast request variables to (int) in all get_sequence/get_next_sequence
callers to prevent SQL injection via $group_query interpolation.

Add intval() to batch ID string construction in api_device_remove_multi,
api_graph_remove_multi, and api_data_source_disable_multi to prevent
injection through deserialized POST arrays.

Convert cli/add_device.php to db_execute_prepared for --ip and
--description arguments that were interpolated directly into SQL.

+++ automation_snmp.php
@@ -617,7 +617,7 @@ function automation_snmp_item_edit() : void {
 		$header_label                     = __('SNMP Options [new]');
 		$automation_snmp_item             = [];
 		$automation_snmp_item['snmp_id']  = grv('id');
-		$automation_snmp_item['sequence'] = get_sequence(0, 'sequence', 'automation_snmp_items', 'snmp_id=' . grv('id'));
+		$automation_snmp_item['sequence'] = get_sequence(0, 'sequence', 'automation_snmp_items', 'snmp_id=' . (int) grv('id'));
 	}
 
 	form_start('automation_snmp.php', 'automation_item_edit');
